+++ lib/github/commands.rb
@@ -2,19 +2,25 @@
 
 require 'subcommand_base'
 require 'github/service'
+require 'github/model/project'
 
 module Kraken
   module GitHub
     class Commands < SubcommandBase
-      desc 'tags', 'List the GitHub tags for a project'
+      desc 'release', 'Creates a release in GitHub for a project'
       option :organization, default: 'jmtrusona'
-      option :repository, default: 'kraken'
-      def tags(client = Kraken::GitHub::Service.new)
-        puts 'Tags'
-        git_tags = client.list_tags(options[:organization], options[:repository])
-        git_tags.each do |git_tag|
-          puts "- #{git_tag.name}"
-        end
+      option :project, default: 'kraken'
+      option :card, default: nil
+      def release(github = Kraken::GitHub::Service.new)
+        release_name = github.release(
+          project: Kraken::GitHub::Project.new(
+            organization: options[:organization],
+            project: options[:project]
+          ),
+          release_notes: 'See CHANGELOG.md for updates',
+          release_card: options[:card]
+        )
+        puts "Created #{release_name}"
       end
     end
   end