+++ spec/trello/commands_spec.rb
@@ -4,12 +4,12 @@ require 'trello/commands'
 
 RSpec.describe Kraken::Trello::Commands do
   context '#cards' do
-    let(:client) { double }
+    let(:trello) { double }
 
-    let(:output) { capture(:stdout) { subject.cards(client) } }
+    let(:output) { capture(:stdout) { subject.cards(trello) } }
 
     it 'lists the cards for the trello board' do
-      expect(client).to receive(:list_cards).and_return([
+      expect(trello).to receive(:list_cards).and_return([
                                                           Kraken::Trello::Card.new(name: 'Build something new')
                                                         ])