+++ spec/trello/service_spec.rb
@@ -3,33 +3,35 @@
 require 'trello/service'
 
 RSpec.describe Kraken::Trello::Service do
-  let(:client) { double }
-  let(:board) { double }
-  let(:list) { double }
-  let(:card) { double }
+  context '#list_cards' do
+    let(:trello_client) { double }
+    let(:board) { double }
+    let(:list) { double }
+    let(:card) { double }
 
-  subject { Kraken::Trello::Service.new }
+    subject { Kraken::Trello::Service.new }
 
-  before { subject.instance_variable_set(:@trello, client) }
+    before { subject.instance_variable_set(:@trello, trello_client) }
 
-  it 'delegates list_cards to the initialized client' do
-    expect(client).to receive(:boards)
-      .and_return([board])
+    it 'lists the cards for the given board' do
+      expect(trello_client).to receive(:boards)
+        .and_return([board])
 
-    expect(board).to receive(:name)
-      .and_return('quesadilla')
+      expect(board).to receive(:name)
+        .and_return('quesadilla')
 
-    expect(board).to receive(:lists)
-      .and_return([list])
+      expect(board).to receive(:lists)
+        .and_return([list])
 
-    expect(list).to receive(:cards)
-      .and_return([card])
+      expect(list).to receive(:cards)
+        .and_return([card])
 
-    expect(card).to receive(:name)
-      .and_return('Build a new thing')
+      expect(card).to receive(:name)
+        .and_return('Build a new thing')
 
-    cards = subject.list_cards('quesadilla')
+      cards = subject.list_cards('quesadilla')
 
-    expect(cards.map(&:name)).to eq(['Build a new thing'])
+      expect(cards.map(&:name)).to eq(['Build a new thing'])
+    end
   end
 end